readin player fixed, new ratinglists

+++ src/main/java/cz/technecium/openkarvinadataservices/domain/PlayerIdentifier.java
@@ -12,6 +12,6 @@ import lombok.Setter;
 @Setter
 public class PlayerIdentifier {
 
-    private long fideId;
-    private long crId;
+    private Integer fideId;
+    private Integer crId;
 }
